moved cache manager to its own file
